requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
azure-cognitiveservices-speech==1.40.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydub==0.25.1